Conflicts:
	electrifyapp/electrifyapp/mainwindow.cpp
	engine/CMakeLists.txt

+++ engine/src/components/oscillators/sin_wave.cpp
@@ -1,6 +1,6 @@
 #include "sin_wave.h"
 
-double SinWave::calculate(const FrameContext& frame) const
+float SinWave::calculate(const FrameContext& frame) const
 {
   return amplitude(frame) * (sin(M_2_PI * frame.timeSeconds() * frequency(frame) + phase(frame)) + 1) / 2;
 }
