+++ engine/src/components/oscillators/square_wave.cpp
@@ -1,6 +1,6 @@
 #include "square_wave.h"
 
-double SquareWave::calculate(const FrameContext& frame) const
+float SquareWave::calculate(const FrameContext& frame) const
 {
   return amplitude(frame) * floor(2 * frame.timeSeconds() * frequency(frame) + phase(frame));
 }
