+++ engine/src/components/time_component.cpp
@@ -1,6 +1,6 @@
 #include "time_component.h"
 
-double TimeComponent::calculate(const FrameContext& frame) const
+float TimeComponent::calculate(const FrameContext& frame) const
 {
   return frame.timeSeconds();
 }
